Move lib/github/notifications to lib/activity
