+++ lib/activity/deployments.js
@@ -1,5 +1,5 @@
-const cache = require('../../cache');
-const { DeploymentStatus } = require('./../../messages/deployment-status');
+const cache = require('../cache');
+const { DeploymentStatus } = require('../messages/deployment-status');
 
 async function deploymentStatus(context, subscription, slack) {
   const deploymentStatusMessage = new DeploymentStatus({
