+++ lib/activity/issues.js
@@ -1,5 +1,5 @@
-const cache = require('../../cache');
-const { Issue } = require('../../messages/issue');
+const cache = require('../cache');
+const { Issue } = require('../messages/issue');
 
 async function updateIssueMessage(messageMetaData, context, subscription, slack) {
   // Fetch updated version of issue to retrieve accurate comment count and body_html
