+++ lib/activity/public.js
@@ -1,4 +1,4 @@
-const { Public } = require('../../messages/public');
+const { Public } = require('../messages/public');
 
 async function publicEvent(context, subscription, slack) {
   const publicMessage = new Public({
