+++ lib/activity/pullRequests.js
@@ -1,6 +1,6 @@
-const { PullRequest } = require('../../messages/pull-request');
+const { PullRequest } = require('../messages/pull-request');
 
-const cache = require('../../cache');
+const cache = require('../cache');
 
 async function pullRequestEvent(context, subscription, slack) {
   const eventType = `${context.event}.${context.payload.action}`;
