+++ lib/activity/push.js
@@ -1,4 +1,4 @@
-const { Push } = require('../../messages/push');
+const { Push } = require('../messages/push');
 
 async function push(context, subscription, slack) {
   if (
