+++ lib/activity/ref.js
@@ -1,4 +1,4 @@
-const { Ref } = require('../../messages/ref');
+const { Ref } = require('../messages/ref');
 
 module.exports = async (context, subscription, slack) => {
   const res = await slack.chat.postMessage(
