+++ lib/activity/review.js
@@ -1,23 +1,21 @@
-const cache = require('../../cache');
-const { Comment } = require('../../messages/comment');
+const cache = require('../cache');
+const { Review } = require('../messages/review');
 
 module.exports = async (context, subscription, slack) => {
-  const { comment, issue, repository } = context.payload;
-
   // Get rendered message
   const message = {
     attachments: [
-      new Comment({ comment, issue, repository }).getRenderedMessage(),
+      new Review({ ...context.payload }).toJSON(),
     ],
   };
 
-  const cacheKey = subscription.cacheKey(`comment#${context.payload.comment.id}`);
+  const cacheKey = subscription.cacheKey(`review#${context.payload.review.id}`);
   const storedMetaData = await cache.get(cacheKey);
 
   if (storedMetaData) {
     const res = await slack.chat.update(storedMetaData.ts, storedMetaData.channel, '', message);
     context.log(res, 'Updated Slack message');
-  } else if (context.payload.action === 'created') {
+  } else if (context.payload.action === 'submitted') {
     const res = await slack.chat.postMessage(subscription.channelId, '', message);
     context.log(res, 'Posted Slack message');
     const messageMetaData = {
