+++ lib/github/index.js
@@ -1,26 +1,7 @@
 const installations = require('./installations');
-const router = require('./router');
-
-const {
-  matchMetaDataStatetoIssueMessage,
-  issueEvent,
-} = require('./notifications/issues');
-
-
-const {
-  pullRequestEvent,
-  onStatus,
-  storePRMapping,
-} = require('./notifications/pullRequests');
-
-const { deploymentStatus } = require('./notifications/deployments');
-const { push } = require('./notifications/push');
-const { publicEvent } = require('./notifications/public');
-const comments = require('./notifications/comments');
-const ref = require('./notifications/ref');
-const review = require('./notifications/review');
 
 const oauth = require('./oauth');
+const activity = require('../activity');
 
 const SignedParams = require('../signed-params');
 
@@ -30,6 +11,8 @@ module.exports = (robot) => {
 
   // set up GitHub oauth
   oauth(robot);
+
+  activity(robot);
 
   const app = robot.route('/github');
 
@@ -57,28 +40,4 @@ module.exports = (robot) => {
       res.redirect(slackAppUrl);
     }
   });
-
-  const route = router(robot);
-
-  robot.on(['issues.opened', 'issues.closed', 'issues.reopened'], route(issueEvent));
-
-  robot.on([
-    'issues.labeled',
-    'issues.unlabeled',
-    'issues.assigned',
-    'issues.unassigned',
-    'issue_comment',
-  ], route(matchMetaDataStatetoIssueMessage));
-
-  robot.on(['pull_request.opened', 'pull_request.closed', 'pull_request.reopened'], route(pullRequestEvent));
-  robot.on(['pull_request.opened', 'pull_request.synchronize'], storePRMapping);
-  robot.on('status', route(onStatus));
-  robot.on('deployment_status', route(deploymentStatus));
-  robot.on('push', route(push));
-  robot.on('public', route(publicEvent));
-
-  robot.on('issue_comment', route(comments));
-
-  robot.on(['create', 'delete'], route(ref));
-  robot.on(['pull_request_review.submitted', 'pull_request_review.edited'], route(review));
 };
